Moved /v1/state/tasks to /v1/tasks Added POST /v1/tasks/restart/<task-name>?replace=<replace>
