+++ src/main/java/org/apache/mesos/scheduler/DefaultScheduler.java
@@ -20,6 +20,7 @@ import org.apache.mesos.specification.ServiceSpecification;
 import org.apache.mesos.state.PersistentOperationRecorder;
 import org.apache.mesos.state.StateStore;
 import org.apache.mesos.state.api.StateResource;
+import org.apache.mesos.scheduler.api.TaskResource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -138,7 +139,8 @@ public class DefaultScheduler implements Scheduler {
         Collection<Object> resources = new ArrayList<>();
         resources.add(new PlanResource(planManager));
         resources.add(new RecoveryResource(recoveryStatusRef));
-        resources.add(new StateResource(stateStore, serviceSpecification.getName()));
+        resources.add(new StateResource(stateStore));
+        resources.add(new TaskResource(stateStore, taskKiller, serviceSpecification.getName()));
         resourcesQueue.put(resources);
     }
