+++ src/main/java/org/apache/mesos/scheduler/recovery/constrain/TimedLaunchConstrainer.java
@@ -3,6 +3,8 @@ package org.apache.mesos.scheduler.recovery.constrain;
 import org.apache.mesos.Protos.Offer.Operation;
 import org.apache.mesos.scheduler.recovery.RecoveryRequirement;
 import org.apache.mesos.scheduler.recovery.RecoveryRequirementUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.time.Duration;
 import java.util.concurrent.atomic.AtomicLong;
@@ -12,6 +14,8 @@ import java.util.concurrent.atomic.AtomicLong;
  * rate-limiting purposes.
  */
 public class TimedLaunchConstrainer implements LaunchConstrainer {
+    private final Logger logger = LoggerFactory.getLogger(getClass());
+
     private AtomicLong lastPermanentRecoveryLaunchMs;
     private Duration minDelay;
 
@@ -38,7 +42,14 @@ public class TimedLaunchConstrainer implements LaunchConstrainer {
     @Override
     public boolean canLaunch(RecoveryRequirement recoveryRequirement) {
         if (RecoveryRequirementUtils.isPermanent(recoveryRequirement)) {
-            return (lastPermanentRecoveryLaunchMs.get() + minDelay.toMillis()) < getCurrentTimeMs();
+            Long timeLeft = lastPermanentRecoveryLaunchMs.get() + minDelay.toMillis() - getCurrentTimeMs();
+            if (timeLeft < 0) {
+                return true;
+            } else {
+                Long secondsLeft = (long) Math.ceil(timeLeft / 1000.0);
+                logger.info("Refusing to launch task for another " + secondsLeft + "s.");
+                return false;
+            }
         } else {
             return true;
         }
